requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
future==0.18.3
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
